+++ packages/router/src/GenericRouter.php
@@ -11,7 +11,6 @@ use Tempest\Http\Mappers\PsrRequestToGenericRequestMapper;
 use Tempest\Http\Request;
 use Tempest\Http\Response;
 use Tempest\Http\Responses\Ok;
-use Tempest\Http\Session\VerifyCsrfMiddleware;
 use Tempest\Router\Exceptions\ControllerActionHadNoReturn;
 use Tempest\Router\Exceptions\MatchedRouteCouldNotBeResolved;
 use Tempest\Router\Routing\Matching\RouteMatcher;
@@ -81,22 +80,6 @@ final readonly class GenericRouter implements Router
                         haystack: $matchedRoute->route->without,
                         strict: true,
                     )) {
-                        return $callable($request);
-                    }
-
-                    // Skip middleware that sets cookies or session values when the route is stateless
-                    if (
-                        $matchedRoute->route->handler->hasAttribute(Stateless::class)
-                        && in_array(
-                            needle: $middlewareClass->getName(),
-                            haystack: [
-                                VerifyCsrfMiddleware::class,
-                                SetCurrentUrlMiddleware::class,
-                                SetCookieMiddleware::class,
-                            ],
-                            strict: true,
-                        )
-                    ) {
                         return $callable($request);
                     }
                 }