+++ packages/upgrade/src/Set/TempestLevelSetList.php
@@ -7,4 +7,5 @@ namespace Tempest\Upgrade\Set;
 final class TempestLevelSetList
 {
     public const string UP_TO_TEMPEST_20 = __DIR__ . '/../../config/sets/level/up-to-tempest-20.php';
+    public const string UP_TO_TEMPEST_28 = __DIR__ . '/../../config/sets/level/up-to-tempest-28.php';
 }